rename segmentation module
