+++ configs/semantic/eval.py
@@ -2,7 +2,7 @@ import matplotlib as mpl
 
 mpl.use("Agg")
 from theseus.opt import Opts
-from theseus.segmentation.pipeline import Pipeline
+from theseus.semantic.pipeline import Pipeline
 
 if __name__ == "__main__":
     opts = Opts().parse_args()
